+++ gantry/core/axis_type_x.cpp
@@ -5,4 +5,4 @@
  *
  * @return an axis type
  */
-extern "C" GantryAxisType get_axis_type() { return gantry_x; }
+extern "C" auto get_axis_type() -> GantryAxisType { return gantry_x; }
