+++ gantry/core/axis_type_y.cpp
@@ -5,4 +5,4 @@
  *
  * @return A node id
  */
-extern "C" GantryAxisType get_axis_type() { return gantry_y; }
+extern "C" auto get_axis_type() -> GantryAxisType { return gantry_y; }
